feat: new /api/packages/<package-name> endpoint & show deprecated message on old /api/js
